misc updates/fixes by Jesse Norell
This patch moves the create_unique_id() function to
misc.c
and makes every place that saves a unique_id use it
(which was
previously done inconsistently in various places). It
also changes
all the use of the status field in messages table to be
through the
following #defines:

#define STATUS_NEW 0
#define STATUS_SEEN 1
#define STATUS_DELETE 2
#define STATUS_PURGE 3
#define STATUS_UNUSED 4
#define STATUS_INSERT 5
#define STATUS_ERROR 6

There are a couple other misc. changes also, namely
changes
the APOP "timestamp" in the pop3 greeting to not
include the
system time and pid (though srand() is still seeded
with those,
but better than giving them away in the clear), put
getopt()
support in pop3d (and added -h help flag), and changed
a few
places to remove messageblks before removing the
message row,
so people adding constraints don't get errors (ie. if
they don't
use ON DELETE CASCADE and remove the messages entry first).

Should apply clean to cvs code around 6/17/03 (probably
a little earlier) to current, 7/18/03. We've been
using it w/o
incident for almost a month now.


git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@802 7b491191-dbf0-0310-aff6-d879d4d69008