merged all 2.0 branch sources with the trunk. 2.0 is now HEAD and 1.2 is in a branch. This makes for an unstable TRUNK, and release (only bugfix) branches.
Quite a lot of effort was needed for this merger. By mistake, I forgot
to add a tag before the branch point for dbmail 2.0 in August, which
caused the fact that a simple branch to trunk merge did not work.. So,
I had to revert to the act of merging file-by-file. Very boring and
error-prone work. Everything seems to be working, but there might be
some errors left.


git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@884 7b491191-dbf0-0310-aff6-d879d4d69008