* fixed a few ACL problems which were found because Mozilla did not function completely with dbmail IMAP
GETACL now also displays the acl of the owner of the mailbox. This ACL is
not coded explicitely in the acl table, but always gives fullrights
to the owner

GETACL, LISTRIGHTS and MYRIGHTS now reply with the name of the mailbox
in quotes, so a mailbox name with whitespace in it is still seen as
a name by the client.


git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@1009 7b491191-dbf0-0310-aff6-d879d4d69008