some changes to mysql functions:
the last_row cache is taken out. It gave us problems, and is not
necessary anymore, because we now only use MySQL >= 4.0.14, which
caches this for itself (this is stuff that should always be done in
the library anyway!)

after a query, mysql_store_result() is always called, to simplify this
function. If we had a function without a result set, the resulting
result set from mysql_store_result() will be NULL, which it would be
anyway.


git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@1049 7b491191-dbf0-0310-aff6-d879d4d69008