Do not assume that a UNION SELECT will be ordered.
There is no guarantee that the result set of a UNION SELECT will be
ordered either in the order of the sub-SELECTs or based on the first
column.

An explicit ORDER BY clause would (empirically)	be too CPU-intensive.
Instead, use the first column of the result set	to determine which
row is which when reading the results.