Clean up usage of 'preauth_capa' and 'capa'.
The separation into the two capability maps is improved, by letting a
session have exactly the capabilities, in each of them, that it's
allowed to use during the pre- and post-authentication phases.  While
'capa' does not change after session initialization, 'preauth_capa'
may change during the pre-authentication phase, for instance when a
plaintext session successfully uses STARTTLS to enable encryption.

The initial setup of capabilities in dbmail_imap_session_new() is
simplified somewhat by this change.

Capability manipulation is isolated to dm_imapsession.c, by adding a
function dbmail_imap_session_encrypted(), that is called if the
session successfully does a STARTTLS, or if it uses TLS to begin with.

The CAPABILITY command now returns the correct capability set for the
current session state in, returning either 'preauth_capa' or 'capa'.