Make support for AUTH=CRAM-MD5 configurable.
Since the use of CRAM-MD5 has been deprecated for quite a while, and
there are clients out there that will try to use it before (or even
without) falling back to PLAIN or LOGIN authentication, a parameter is
added to the IMAP section of dbmail.conf.

While here, improve the checks for the requested authentication method
in _ic_authenticate(), so that unsupported methods will fail early,
without the server parsing (and asking for, if missing) the parameters
to a method that is disabled in 'preauth_capa'.